core(cli): merge cli.rs into cmd
